% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_samples.R
\name{compare_samples}
\alias{compare_samples}
\title{compare_samples}
\usage{
compare_samples(input1, input2)
}
\arguments{
\item{input1}{first output}

\item{input2}{second output}
}
\value{
A data.frame of seq.ID (columns) and sample.ID (rows) with either relative or absolute abundance of sequences.
}
\description{
compare_samples
}
