% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_seqs.R
\name{extract_seqs}
\alias{extract_seqs}
\title{Function to extract sequences from symportal output}
\usage{
extract_seqs(
  folder,
  metadata = NULL,
  type = "absolute",
  factors = NULL,
  onlyprofile = FALSE,
  remove_zero = TRUE,
  arrange = TRUE,
  seed = NULL,
  ...
)
}
\arguments{
\item{folder}{location of the root Symportal output}

\item{metadata}{location of a metadata file in csv, must contain a column called sample_name with matches}

\item{type}{returns either "relative" or "absolute"}

\item{factors}{option to include a vector list of column names, e.g: c("location", "latitude") to not import every column in a metadata file, if left NULL then will import all columns}

\item{onlyprofile}{TRUE/FALSE return only profiles}

\item{remove_zero}{drop empty rows / columns}

\item{arrange}{arrange by seq.ID alphanumerically, TRUE by default}

\item{seed}{value for set.seed}
}
\value{
A data.frame of seq.ID (columns) and sample_name (rows) with either relative or absolute abundance of sequences.
}
\description{
extract sequences from symportal output
}
