% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_seqs.R
\name{plot_seqs}
\alias{plot_seqs}
\title{compare_samples}
\usage{
plot_seqs(
  input,
  folder,
  type = "ggplot",
  cluster = "none",
  nrow = NULL,
  facet = NULL,
  seq.order = TRUE,
  ...
)
}
\arguments{
\item{input1}{first output}

\item{input2}{second output}
}
\value{
A data.frame of seq.ID (columns) and sample.ID (rows) with either relative or absolute abundance of sequences.
}
\description{
compare_samples
}
